Tweak texture load order for minimum visual latency
So as it turns out Unity is only willing to issue one texture load per
frame. This means that the order in which we issue texture loads
basically controls when they complete. This doesn't really matter when
switching scenes, but it is noticeable when tabbing between planets
and/or vessels in the map view and tracking station.

This commit tweaks the load order for scaled bodies so that it goes like
this:
* important textures: _ColorMap, _NormalMap, _HeightMap, _BumpMap
  (in that order)
* all other scaled body textures
* all other terrain body textures, if needed by the scaled body.

This minimizes the latency between a load starting and the body looking
correct. It is still noticeable if you're looking for it but is now even
less noticeable overall.