Use KSPTextureLoader to load textures
I have extracted the contents of #85 out to a separate library and then
iterated on it a bunch more to get even better performance where
possible.

Here's how this works from Parallax's side: you load a texture by
calling `TextureLoader.LoadTexture<T>`. This gives you back a
TextureHandle, which is basically a reference-counted handle for a
texture. You can wait for this to complete by doing
`yield return handle` or you can block on it immediately by calling
`handle.GetTexture()`. When you are done with the texture you call
`handle.Dispose()`, which will take care of destroying the texture when
all references are gone.

Behind the scenes KSPTextureLoader will transparently take care of:
* Loading textures from asset bundles when configured.
* Sharing textures between all mods that request the same texture.
* All the necessary cursed tricks you need to do to actually load a
  texture in unity with minimal overhead.

The nice bit here is that this means that both parallax and kopernicus
will be able to share loaded textures, once both of my PRs land.