Disable Kopernicus ScaledOnDemandComponent for parallax planets
With both of these enabled there is a race condition between parallax
and kopernicus to set certain properties on the shader. Whichever one
finishes last will set the _MainTex and _BumpMap properties on the
shader. If the textures are different between kopernicus and parallax
then this will mean that planets sometimes work correctly and sometimes
work wrong.

This fixes the issue by destroying kopernicus' ScaledOnDemandComponent
for any planet whose scaled space object is managed by parallax.